trying anything at this point

+++ scripts/clean.mjs
@@ -4,13 +4,13 @@ import { deleteAsync } from "del";
 import arg from "arg";
 import path from "path";
 
-const turboCachePaths = ["./**/.turbo"];
+const turboCachePaths = ["./apps/**/.turbo"];
 const buildOuputPaths = [
   ".husky",
   ".parcel-cache",
-  "browser-extension/dist",
-  "graphql/dist",
-  "ui/.next",
+  "apps/browser-extension/dist",
+  "apps/graphql/dist",
+  "apps/ui/.next",
 ];
 
 const args = arg({
@@ -20,19 +20,21 @@ const args = arg({
   "-d": "--dryrun",
 });
 
-const removedPaths = await deleteAsync(
-  args["--all"] ? [...turboCachePaths, ...buildOuputPaths] : buildOuputPaths,
-  {
-    dryRun: args["--dryrun"],
+if (!process.env.CI) {
+  const removedPaths = await deleteAsync(
+    args["--all"] ? [...turboCachePaths, ...buildOuputPaths] : buildOuputPaths,
+    {
+      dryRun: args["--dryrun"],
+    }
+  );
+
+  if (args["--dryrun"]) {
+    console.log("Would-be deleted directories:\n");
+  } else {
+    console.log("Deleted directories:\n");
   }
-);
 
-if (args["--dryrun"]) {
-  console.log("Would-be deleted directories:\n");
-} else {
-  console.log("Deleted directories:\n");
-}
-
-for (const removed of removedPaths) {
-  console.log("♻️ ", path.relative(process.cwd(), removed));
+  for (const removed of removedPaths) {
+    console.log("♻️ ", path.relative(process.cwd(), removed));
+  }
 }
